+++ src/app/cart/services/cart.service.ts
@@ -3,8 +3,6 @@ import {Injectable} from '@angular/core';
 @Injectable({
   providedIn: 'root'
 })
-export class CartServiceService {
+export class CartService {
 
-  constructor() {
-  }
 }