+++ src/app/products/models/product.model.ts
@@ -1,9 +1,0 @@
-import {Category} from './category.enum';
-
-export interface ProductModel {
-  name: string;
-  description: string;
-  price: number;
-  category: Category;
-  isAvailable: boolean;
-}